fix: move tiling functions from raster to tiling
